requirements.txt
Django>=5.0
djangorestframework
pandas
requests
python-dotenv
numpy
scipy
geopy
polyline
